#804 Move test-only invoker accessor away from API
